[Security Solution] Batch prebuilt rule installation (#214441)
**This is a follow-up to:https://github.com/elastic/kibana/pull/211045**

## Summary

This PR removes inefficiencies in prebuilt rule installation memory
consumption.

### Before

In the worst-case scenario:

1. All currently installed prebuilt rules were fully loaded into memory.
2. All latest rule versions from the rule packages were fully loaded
into memory.
3. All base rule versions were pulled into memory, even though they were
not used.
4. The algorithm then checked which rules could be installed and
installed them all at once.

### After

1. Pull only aggregated information about installed rules (only
`rule_id` and `versions`).
2. Pull the same lightweight aggregated info about the latest rules in
the package.
3. Using the collected `rule_id`s, fetch rule assets and install them in
small batches of up to 100 rules.

(cherry picked from commit 6d9fc21db92dd5e02d930971a5a1cf32a97386f0)
